requirements.txt
dvc
evidently
gdown
joblib
mlflow==2.8.1
numpy
pandas
python-box
PyYAML
setuptools
tqdm
typeguard
types-PyYAML
nltk==3.8.1
lightgbm==4.5.0
seaborn
dagshub
-e .
